Do not attempt to do avahi stuff on non-linux
Prior to this commit, avahi would be applied to any node including the
avahi class.  This is great until you have a mac check in and there is
not avahi package in any way.  So, I have now constrained the module to
only run on Linux clients, even if included elsewhere.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>